+++ src/cli.rs
@@ -348,10 +348,11 @@ pub fn read_manifest(path: &Path) -> Result<Manifest> {
         .map_err(|cause| Error::chain(format!("Cannot open manifest {:?}.", path), cause))?;
     let partial: PartialManifest = serde_yaml::from_reader(file)
         .map_err(|cause| Error::chain(format!("Syntax error in manifest {:?}.", path), cause))?;
-    let manifest = partial
+    partial
+        .prefix_paths(path.parent().unwrap())
+        .map_err(|cause| Error::chain(format!("Error in manifest prefixing {:?}.", path), cause))?
         .validate()
-        .map_err(|cause| Error::chain(format!("Error in manifest {:?}.", path), cause))?;
-    manifest.prefix_paths(path.parent().unwrap())
+        .map_err(|cause| Error::chain(format!("Error in manifest {:?}.", path), cause))
 }
 
 /// Load a configuration by traversing a directory hierarchy upwards.